+++ src/lib.rs
@@ -13,11 +13,20 @@ pub mod modules;
 pub mod preprocessors;
 
 pub fn new_greco_rt_builder() -> EsRuntimeBuilder {
-    let mut rt_builder = EsRuntimeBuilder::new();
+    new_greco_rt_builder2(true, true, true)
+}
 
-    rt_builder = modules::init(rt_builder);
-    rt_builder = features::init(rt_builder);
-    rt_builder = preprocessors::init(rt_builder);
+pub fn new_greco_rt_builder2(preprocs: bool, features: bool, modules: bool) -> EsRuntimeBuilder {
+    let mut rt_builder = EsRuntimeBuilder::new();
+    if modules {
+        rt_builder = modules::init(rt_builder);
+    }
+    if features {
+        rt_builder = features::init(rt_builder);
+    }
+    if preprocs {
+        rt_builder = preprocessors::init(rt_builder);
+    }
 
     rt_builder
 }
@@ -26,8 +35,10 @@ pub fn new_greco_rt_builder() -> EsRuntimeBuilder {
 pub mod tests {
     use crate::new_greco_rt_builder;
 
+    use backtrace::Backtrace;
     use log::LevelFilter;
     use quickjs_runtime::esruntime::EsRuntime;
+    use std::panic;
     use std::sync::Arc;
 
     #[test]
@@ -37,6 +48,15 @@ pub mod tests {
     }
 
     pub fn init_test_greco_rt() -> Arc<EsRuntime> {
+        panic::set_hook(Box::new(|panic_info| {
+            let backtrace = Backtrace::new();
+            log::error!(
+                "thread panic occurred: {}\nbacktrace: {:?}",
+                panic_info,
+                backtrace
+            );
+        }));
+
         simple_logging::log_to_file("greco_rt.log", LevelFilter::Trace)
             .ok()
             .unwrap();