+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 0.0.2 (work in progress)
 
+* ifdef preprocessor
+* macro preprocessor
+
 0.0.1
 
 * started tracking changes